refactor: refactor app struct
