+++ src/app/app.component.html
@@ -1,6 +1,10 @@
 <header>
   <div class="wrapper header-block">
     <span class="logo">TvViewer</span>
+    <nav *ngIf="isLoggedIn$ | async">
+      <a [routerLink]="['channels']">Channels</a>
+      <a [routerLink]="['favorite']">Favorite</a>
+    </nav>
     <mat-icon (click)="openRegistrModalWindow()">power_settings_new</mat-icon>
   </div>
 </header>
