+++ src/app/modules/channel-info/components/tvshow/tvshow.component.ts
@@ -11,8 +11,8 @@ import {
 import * as moment from 'moment';
 import { Moment } from 'moment';
 import { BehaviorSubject } from 'rxjs';
-import { TvshowModel } from '../../../../interfaces/tvshow.model';
-import { ProcessingService } from '../../../../services/processing.service';
+import { TvshowModel } from '../../../../shared/interfaces/tvshow.model';
+import { ProcessingService } from '../../../../shared/services/processing.service';
 
 @Component({
   selector: 'app-tvshow',
