+++ src/app/modules/favorite/favorite.component.html
@@ -1,0 +1,11 @@
+<div class="wrapper">
+  <div class="channels">
+    <h1>Favorite</h1>
+<!--    <mat-spinner *ngIf="!filtredChannelList && filtredChannelList !== null"></mat-spinner>-->
+<!--    <div *ngIf="filtredChannelList === null">Channels not found. Please, change channel category</div>-->
+
+<!--    <div *ngFor="let channel of filtredChannelList" >-->
+<!--      <app-channel-card [info]="channel"></app-channel-card>-->
+<!--    </div>-->
+  </div>
+</div>
