+++ src/app/modules/favorite/favorite.module.ts
@@ -1,0 +1,9 @@
+import {NgModule} from "@angular/core";
+import {CommonModule} from "@angular/common";
+import {FavoriteComponent} from "./favorite.component";
+
+@NgModule({
+  declarations: [FavoriteComponent],
+  imports: [CommonModule],
+})
+export class FavoriteModule {}
