+++ src/app/shared/components/login-form/validators/email.validator.ts
