+++ src/app/shared/interfaces/auth-token.model.ts
