+++ src/app/shared/interfaces/categories.model.ts
