+++ src/app/shared/interfaces/channel.model.ts
