+++ src/app/shared/interfaces/favorite-channels-list.interface.ts
