+++ src/app/shared/interfaces/tvshow-list.model.ts
