+++ src/app/shared/interfaces/tvshow-wripper.model.ts
