+++ src/app/shared/interfaces/video-info.model.ts
