+++ src/app/shared/interfaces/video-wripper.model.ts
