+++ src/app/shared/services/favorite-channel.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import {FavoriteChannelsListInterface} from "../interfaces/favorite-channels-list.interface";
+import { FavoriteChannelsListInterface } from '../interfaces/favorite-channels-list.interface';
 
 @Injectable({
   providedIn: 'root',
