+++ src/app/store/selectors.ts
@@ -1,4 +1,4 @@
-import {createFeatureSelector, createSelector, Store} from '@ngrx/store';
+import { createFeatureSelector, createSelector, Store } from '@ngrx/store';
 import { AppStateInterface } from './types/app-state.inteface';
 import { StoreStateInterface } from './types/store-state.interface';
