+++ src/app/store/types/store-state.interface.ts
@@ -1,4 +1,4 @@
-import { FavoriteChannelInterface } from '../../interfaces/favorite-channels-list.interface';
+import { FavoriteChannelInterface } from '../../shared/interfaces/favorite-channels-list.interface';
 
 export interface StoreStateInterface {
   favoriteChannels: FavoriteChannelInterface[];
